refactor: improve project structure and code readability v3

+++ hibernate-blank/src/main/java/pl/edu/agh/mwo/hibernate/filealbummanager/repository/UserRepository.java
@@ -4,7 +4,7 @@ import org.hibernate.Session;
 import org.hibernate.Transaction;
 import org.hibernate.query.Query;
 import pl.edu.agh.mwo.hibernate.filealbummanager.entity.User;
-import pl.edu.agh.mwo.hibernate.filealbummanager.ui.Messages;
+import pl.edu.agh.mwo.hibernate.filealbummanager.ui.account.AccountMessages;
 
 import java.util.List;
 
@@ -18,18 +18,23 @@ public class UserRepository {
 
     public User getUserFromDatabase(String userName) {
         Query<User> query = session.createQuery("FROM User u WHERE u.name = :name", User.class);
+
         query.setParameter("name", userName);
+
         return query.uniqueResult();
     }
 
     public User getUserFromDatabase(int userId) {
         Query<User> query = session.createQuery("FROM User u WHERE u.id = :id", User.class);
+
         query.setParameter("id", userId);
+
         return query.uniqueResult();
     }
 
     public List<User> getUsersFromDatabase() {
         Query<User> query = session.createQuery("FROM User", User.class);
+
         return query.list();
     }
 
@@ -40,20 +45,43 @@ public class UserRepository {
     public void addUser(String userName) {
         User user = new User();
         user.setName(userName);
+
         Transaction transaction = session.beginTransaction();
-        session.save(user);
-        transaction.commit();
+
+        try {
+            session.save(user);
+            transaction.commit();
+        } catch (Exception e) {
+            if (transaction.isActive()) {
+                transaction.rollback();
+            }
+            throw e;
+        }
     }
 
     public void deleteUser(User user) {
+        if (user == null) {
+            return;
+        }
+
         Transaction transaction = session.beginTransaction();
-        session.delete(user);
-        transaction.commit();
+
+        try {
+            session.delete(user);
+            transaction.commit();
+        } catch (Exception e) {
+            if (transaction.isActive()) {
+                transaction.rollback();
+            }
+            throw e;
+        }
     }
 
     public void printUsers() {
         List<User> users = getUsersFromDatabase();
-        System.out.println(Messages.USERS_HEADER);
+
+        System.out.println(AccountMessages.USERS_HEADER);
+
         for (User user : users) {
             System.out.println(user);
         }
